feat(web): discovery mode switcher (Search / Browse / Swipe)
First slice of #340: a `role="tablist"` segmented control above the
main column routes between freeform Search (existing want-list
editor + results), inline Browse (BrowseModal's UI promoted into
the page), and a Swipe placeholder.

BrowseModal state + effects move into a new `useBrowseController`
hook so the modal and the inline panel share one source of truth
without duplicating fetch / cache / reset behavior. BrowseModal
becomes a thin Radix Dialog wrapper around the new BrowsePanel.

Four new App-level tests cover the default Search mode, the inline
Browse panel, the Swipe placeholder copy + link, and round-tripping
back to Search without losing the editor's input. All 16 existing
BrowseModal tests still pass against the refactored wrapper.
