+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Versions follow [Semantic Versioning](https://semver.org/).
 
 ### Added
 
+- Web: **Discovery mode switcher — Search / Browse / Swipe tabs above the main column** ([#340](https://github.com/mgzwarrior/mgz-pkmn/issues/340)). First slice of the discovery-modes epic, scoped down to a single iteration: a `role="tablist"` segmented control over the main content area routes between freeform Search (the existing want-list editor + results), inline Browse (the BrowseModal's set-list + set-detail UI promoted into the page, sharing all state with the modal so the header chip still works), and a Swipe placeholder section that links back to issue [#340](https://github.com/mgzwarrior/mgz-pkmn/issues/340) for follow-up. BrowseModal's state + effects were lifted into a new `useBrowseController` hook so the modal and the inline panel share one source of truth without duplicating fetch / cache / reset behavior; the existing 16 BrowseModal tests still pass against the refactored wrapper. Four new App-level tests in `web/src/App.test.tsx` cover the default Search mode, the Browse panel rendering inline, the Swipe placeholder copy + link, and round-tripping back to Search without losing the editor's input.
+
 - Web: **SPA sign-in UI — header chip, provider picker, signed-in state** ([#411](https://github.com/mgzwarrior/mgz-pkmn/issues/411)). Fifth slice of the [#61](https://github.com/mgzwarrior/mgz-pkmn/issues/61) hosted-demo auth epic, per [ADR-0019](docs/adr/0019-hosted-demo-identity-and-auth.md). Turns the auth backend ([#407](https://github.com/mgzwarrior/mgz-pkmn/issues/407) / [#408](https://github.com/mgzwarrior/mgz-pkmn/issues/408) / [#409](https://github.com/mgzwarrior/mgz-pkmn/issues/409) / [#410](https://github.com/mgzwarrior/mgz-pkmn/issues/410)) into something a visitor can actually use. New `web/src/components/SignInChip.tsx` mounts in the header next to the theme toggle as an icon-only control (the sign-in glyph for anonymous sessions, a circular initials avatar for signed-in ones) so the header stays uncluttered as the nav row grows. Anonymous sessions open a Radix dialog with three provider options — GitHub (anchor to `/api/v1/auth/github/login`), Google (anchor to `/api/v1/auth/google/login`), and a magic-link button that expands inline into an email field and POSTs to `/api/v1/auth/magic/request`, then shows a "check your inbox" confirmation (no enumeration leak: the response shape is identical for known and unknown addresses by design of the backend's 202). Signed-in sessions open a Radix dropdown that surfaces the user's `display_name` and email at the top of the menu plus a "Sign out" item, which POSTs to `/api/v1/auth/logout` and flips the chip back to anonymous. State is backed by a new `useAuth` hook in `web/src/hooks/useAuth.ts` that polls `GET /api/v1/me` once on mount and maps 204 No Content to `null` so the chip's signed-out path doesn't have to know about the status-code split — when a provider's OAuth callback 302s back to `/`, the hook's mount-time fetch re-runs and the chip flips to the signed-in shape without the SPA having to thread anything through the URL. New API-client helpers (`fetchMe`, `logout`, `requestMagicLink`) in `web/src/api/client.ts` keep the network surface in one place. Pinned by six new component tests in `web/src/components/SignInChip.test.tsx` covering the anonymous chip + picker shape, the OAuth anchor hrefs, the magic-link submit + confirmation + error paths, the signed-in dropdown's Sign-out flow, and the initials fallback when `display_name` is null; the existing `a11y.test.tsx` scan adds the anonymous chip and the open picker (with the magic-link form expanded) so axe enforces no violations in any of the three new dialog states.
 
 - API: **Google OAuth sign-in** ([#410](https://github.com/mgzwarrior/mgz-pkmn/issues/410)). Fourth slice of the [#61](https://github.com/mgzwarrior/mgz-pkmn/issues/61) hosted-demo auth epic, per [ADR-0019](docs/adr/0019-hosted-demo-identity-and-auth.md). New `api/auth/google.py` registers an Authlib OAuth client against Google's OIDC discovery doc (`https://accounts.google.com/.well-known/openid-configuration`) via two env vars (`MGZ_PKMN_GOOGLE_CLIENT_ID` + `MGZ_PKMN_GOOGLE_CLIENT_SECRET`) and exposes `GET /api/v1/auth/google/login` → authorize redirect, `GET /api/v1/auth/google/callback` → ID-token exchange + verified-email read + upsert against the `users` table + session-cookie issue + 302 back to `/`. Scope is `openid email profile` — no Gmail or Drive access. Account-merge follows the same ADR-0019 first-set-wins contract as the GitHub + magic-link providers: the row is keyed on the verified email (an `email` claim without `email_verified=true` is rejected as `no_verified_email`), `display_name` is only populated when the existing row has none, and `email_verified_at` is stamped on the first sign-in that brings a verified email through. `users.name` is namespaced under `google:` followed by Google's stable `sub` identifier, keeping it domain-separated from the `gh:` and `magic:` prefixes minted by the other providers. Both routes 404 cleanly when `MGZ_PKMN_AUTH_ENABLED` is off, and 503 with a friendly message when the env vars aren't set, instead of leaking an Authlib traceback. `render.yaml` declares the two new env vars under the `mgz-pkmn-api` service with `sync: false` so the deploy knows they're required without storing the values in git. Pinned by 10 new tests in `tests/test_auth_google.py` covering the login gate (auth-off → 404 with Starlette-default body, env-missing → 503), the callback error branches (Authlib `MismatchingStateError` → 400, missing-or-unverified email → 400, empty `sub` → 400), the upsert paths (fresh signup creates a `google:`-prefixed row; existing row with a `display_name` keeps it; existing row without one is filled; concurrent-signup race recovers via `IntegrityError` rollback + re-read instead of 500), and the session-cookie acceptance criterion (the callback's cookie authenticates a subsequent `/api/v1/me` call to 200 with the new user's payload).
